refactor(hero-section.tsx): refactored hero section and google login
