+++ src/utils/hero/shuffle.ts
@@ -1,0 +1,18 @@
+import { heroSectionSquareData } from "@lib/constants";
+
+export const shuffle = (array: (typeof heroSectionSquareData)[0][]) => {
+  let currentIndex = array.length,
+    randomIndex;
+
+  while (currentIndex != 0) {
+    randomIndex = Math.floor(Math.random() * currentIndex);
+    currentIndex--;
+
+    [array[currentIndex], array[randomIndex]] = [
+      array[randomIndex],
+      array[currentIndex],
+    ];
+  }
+
+  return array;
+};
